requirements.txt
black
pylint
pytest
pytest-cov
python-dotenv==0.21.1
requests==2.28.2
google-api-python-client==2.80.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
pynecone==0.1.18